Incorporate navigation bar template
- Copy work in the project at
https://github.com/alanecode/skeleton-css-menu into this repo,
containing css and other resources for a mobile first navitation bar.

- Format makes use of the Skeleton CSS boilerplate, but uses CSS rather
than jquery (as is done on the Skeleton demo page) to achieve mobile
first navigation.

- Font Awesome fonts added to public/fonts to provide symbols for
buttons.

- New .ejs template containing partial for navigation bar added.

- Former header.ejs and footer.ejs templates refactored, and new
head.ejs template introduced. This is to avoid the situation in which,
e.g. <html> tag was opened in header.ejs and closed in footer.ejs. Now
page templates such as index.ejs constitute complete html documents,
except for the partial templates (head.ejs, header.ejs...) which are
included at the appropriate points. This is to aid in code readability
and reusability across page templates, and follows the example given for
using ejs templates with express at
https://scotch.io/tutorials/use-ejs-to-template-your-node-application