Correct order of navigation menu items
Problem of menu items being swapped around was caused by the li elements
in the list of navigation options being floated right (revering the
order).

The trick was to float the enclosing ul element right, with the li items
floated left. Then, a margin-left was given to the 5th element to push
the bottom row to the right, giving it the appearance of being aligned
right.